backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
mysqlclient>=2.2
python-dotenv>=1.0
tensorflow>=2.16
pandas>=2.2
scikit-learn>=1.3.0
joblib>=1.2.0
numpy>=1.24.0
bcrypt>=4.0.0
PyJWT>=2.8.0